+++ test/test_helper.exs
@@ -2,5 +2,4 @@ if System.get_env("CI") == "true" and function_exported?(Code, :put_compiler_opt
   Code.put_compiler_option(:warnings_as_errors, true)
 end
 
-Finch.start_link(name: Mxpanel.HTTPClient)
 ExUnit.start()